requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.138
botocore==1.34.138
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cover-agent @ git+https://github.com/ankitchoudhary2209/cover-agent.git@619b02375a9cc84fdea1e2938204dbac31ababa1
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
dunamai==1.21.2
dynaconf==3.2.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-aiplatform==1.57.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.4
jmespath==1.0.1
litellm @ git+https://github.com/mrT23/litellm.git@61556f0c9803674339554c8252001a67bc112768
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==1.35.9
packaging==24.1
platformdirs==4.2.2
poetry-dynamic-versioning==1.4.0
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.0
pydantic_core==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==70.2.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.5
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wandb==0.17.3
yarl==1.9.4
zipp==3.19.2
